+++ packages/api-reference/src/components/Content/Operation/ExampleRequest.vue
@@ -76,7 +76,7 @@ const generateSnippet = async (): Promise<string> => {
   // Generate a request object
   const request = getHarRequest(
     {
-      url: getUrlFromServerState(serverState) ?? window.location.origin,
+      url: getUrlFromServerState(serverState),
     },
     getRequestFromOperation(
       props.operation,