+++ packages/nuxt/.eslintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ['@nuxt/eslint-config', '../../.eslintrc.cjs'],
+}